Recognition-Algorithms/Face Recognition using FaceNet/requirements.txt
numpy==1.16.4
Pillow==6.1.0
matplotlib==3.1.0
opencv-contrib-python==4.1.0.25
scipy==1.2.1
scikit-learn==0.21.2
tensorflow==1.14.0
Keras==2.2.4